2024/08/02 后端开发：课费预支付功能对上课签到/撤销签到的影响，以及处理。 1.上课签到：正常情况下，一旦签到，同时就对《课费表》进行新数据插入操作。           在有了课费预支付功能后，因为预支付时，已经对《课费表》做了新规插入操作，           所以不能再第二次执行重复的新规操作了，所以要追加该lesson_id是否时预先支付课的判断。 2.撤销签到：正常情况下，一旦撤销，同时也对《课费表》进行相应的删除操作。           在有了课费预支付功能后，因为预支付时，因为《课程表》与《课费表》直接有外建约束的数据完整性，           不能对《课费表》执行操作，所以要追加该lesson_id是否时预先支付课的判断。

+++ 01FlutterApps/middle_layer_springboot/knpiano_java/src/main/java/com/liu/springboot04web/dao/Kn01L002LsnDao.java
@@ -2,6 +2,7 @@ package com.liu.springboot04web.dao;
 
 import com.liu.springboot04web.bean.Kn01L002LsnBean;
 import com.liu.springboot04web.bean.Kn02F002FeeBean;
+import com.liu.springboot04web.bean.Kn02F003LsnFeeAdvcPayBean;
 import com.liu.springboot04web.bean.Kn02F004PayBean;
 import com.liu.springboot04web.bean.Kn03D004StuDocBean;
 import com.liu.springboot04web.constant.KNConstant;
@@ -27,6 +28,8 @@ public class Kn01L002LsnDao {
     private Kn02F002FeeDao    kn02F002FeeDao;
     @Autowired
     private Kn02F004PayDao    kn02f004PayDao;
+    @Autowired
+    Kn02F003LsnFeeAdvcPayDao  kn02F003LsnFeeAdvcPayDao;
 
     public List<Kn01L002LsnBean> getInfoList(String year) {
         List<Kn01L002LsnBean> list =knLsn001Mapper.getInfoList(year);
@@ -76,8 +79,21 @@ public class Kn01L002LsnDao {
             knLsn001Bean.setScanQrDate(new Date());
             save(knLsn001Bean);
 
-            // 对签到的课程执行课费计算 
-            addNewLsnFee(knLsn001Bean);
+            // 如果课费预支付表里已经有了该课的预支付记录，就表示不必再往课费表里进行插入操作，否则会发生主键冲突
+            String lessonId = knLsn001Bean.getLessonId();
+            Kn02F003LsnFeeAdvcPayBean advcPaidBean = 
+                        kn02F003LsnFeeAdvcPayDao.getAdvcFeePaidyInfoByIds(lessonId, null, null);
+            if (knLsn001Bean.getLessonType() == 1 && advcPaidBean != null) {
+                // 对课费预支付表里的该当lesson_id的advc_flg值做更新操作（advc_flg值，从0更新为1）
+                advcPaidBean.setAdvcFlg(1);
+                kn02F003LsnFeeAdvcPayDao.update(advcPaidBean);
+
+                // 对课程表不做任何操作（月计划课程在课费预支付表里存在的话，不能再往课费表里执行插入操作）。
+                // 重复的lesson_id会导致主键冲突的数据库错误。
+            } else {
+                // 对签到的课程执行课费计算 
+                addNewLsnFee(knLsn001Bean);
+            }
         }
     }
 
@@ -167,9 +183,23 @@ public class Kn01L002LsnDao {
         List<Kn02F002FeeBean> searchResults = kn02F002FeeDao.searchLsnFee(condition);
         Kn02F002FeeBean feeBean = searchResults.get(0);
 
-        // 删除当日的课费计算记录
-        kn02F002FeeDao.delete(feeBean.getLsnFeeId(), feeBean.getLessonId());
+        // 如果课费预支付表里已经有了该课的预支付记录，就表示不必再往课费表里进行插入操作，否则会发生主键冲突
+        String lessonId = knLsn001Bean.getLessonId();
+        Kn02F003LsnFeeAdvcPayBean advcPaidBean = 
+                    kn02F003LsnFeeAdvcPayDao.getAdvcFeePaidyInfoByIds(lessonId, null, null);
+        if (knLsn001Bean.getLessonType() == 1 && advcPaidBean != null) {
+            // 对课费预支付表里的该当lesson_id的advc_flg值做更新操作（advc_flg值，从1还原为0）
+            advcPaidBean.setAdvcFlg(0);
+            kn02F003LsnFeeAdvcPayDao.update(advcPaidBean);
+
+            // 因为课费预支付操作就有了《课程表》与《课费表》直接有外建约束的数据完整性，，此处不能执行课费表的删除操作。
+            // 对《课费表》当日的课费记录不做删除处理
+        } else {
+            // 对《课费表》当日的课费记录做删除处理
+            kn02F002FeeDao.delete(feeBean.getLsnFeeId(), feeBean.getLessonId());
+        }
     }
+
     // 撤销签到
     public void restoreSignedLsn(String lessonId) {
         knLsn001Mapper.restoreSignedLsn(lessonId);
